+++ src/controllers/PhotoWizard/PhotoWizardViewController.m
@@ -42,8 +42,8 @@ static NSString *const LOCATIONSUBSCRIBERID=@"PhotoWizard";
 @property (nonatomic, strong) IBOutlet UIView *headerView;
 @property (nonatomic, strong) IBOutlet UIView *footerView;
 @property (nonatomic, strong) LayoutBox *pageContainer;
-@property (nonatomic, assign) int activePage;
-@property (nonatomic, assign) int maxVisitedPage;
+@property (nonatomic, assign) NSInteger activePage;
+@property (nonatomic, assign) NSInteger maxVisitedPage;
 @property (nonatomic, strong) NSMutableArray *viewArray;
 @property (nonatomic, strong) IBOutlet UIToolbar *modalToolBar;
 @property (nonatomic, strong) IBOutlet UIBarButtonItem *cancelViewButton;
@@ -129,7 +129,7 @@ static NSString *const LOCATIONSUBSCRIBERID=@"PhotoWizard";
 -(void)initCategoryView:(PhotoWizardViewState)state;
 -(void)updateCategoryView;
 -(void)resetCategoryView;
--(void)didSelectCategoryFromMenu:(int)index;
+-(void)didSelectCategoryFromMenu:(NSInteger)index;
 
 -(void)initDescriptionView:(PhotoWizardViewState)state;
 -(void)updateDescriptionView;
@@ -1177,7 +1177,7 @@ static NSString *const LOCATIONSUBSCRIBERID=@"PhotoWizard";
 }
 
 
--(void)didSelectCategoryFromMenu:(int)index{
+-(void)didSelectCategoryFromMenu:(NSInteger)index{
 	
 	PhotoCategoryVO *vo=_activePickerDataSource[index];
